MNTOR-2374 - remove standalone subscription page and replace with reusable component, and show errors instead of redirecting to /
