+++ src/app/(proper_react)/redesign/(authenticated)/user/dashboard/fix/data-broker-profiles/welcome-to-premium/WelcomeToPremiumView.tsx
@@ -5,6 +5,7 @@
 import styles from "./welcomeToPremium.module.scss";
 import { getL10n } from "../../../../../../../../functions/server/l10n";
 import { PercentageChart } from "../../../../../../../../components/client/PercentageChart";
+import { SubscriptionCheck } from "../../../../../../../../components/client/SubscriptionCheck";
 import {
   getDashboardSummary,
   getExposureReduction,
@@ -84,6 +85,9 @@ export function WelcomeToPremiumView(props: Props) {
         </div>
         <div className={styles.chart}>
           <PercentageChart exposureReduction={exposureReduction} />
+        </div>
+        <div>
+          <SubscriptionCheck />
         </div>
       </div>
     </FixView>
