--- src/app/(proper_react)/redesign/(authenticated)/user/dashboard/subscribed/page.tsx
@@ -1,57 +1,0 @@
-/* This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
-
-"use client";
-
-import { useSession } from "next-auth/react";
-import { useEffect } from "react";
-import { useRouter } from "next/navigation";
-
-import { hasPremium } from "../../../../../../functions/universal/user";
-import { captureException } from "@sentry/browser";
-import { useL10n } from "../../../../../../hooks/l10n";
-import { captureMessage } from "@sentry/nextjs";
-
-/**
- * Client-side page to update session info.
- *
- * Next-Auth does not have a simple way to do this purely from the server-side, so we
- * use this page to check and redirect appropriately.
- *
- * NOTE: this does not replace doing server-side `hasPremium` checks! This is just
- * a convenience so users do not need to sign out and back in to refresh their session
- * after subscribing.
- */
-export default function Subscribed() {
-  const l10n = useL10n();
-  const { update } = useSession();
-  const router = useRouter();
-
-  useEffect(() => {
-    async function updateSession() {
-      try {
-        const result = await update();
-        if (hasPremium(result?.user)) {
-          router.replace(
-            `/redesign/user/dashboard/fix/data-broker-profiles/welcome-to-premium`,
-          );
-        } else {
-          const message = "No subscription found";
-          captureMessage(message);
-          console.error(message);
-          router.replace("/redesign/user/dashboard");
-        }
-      } catch (ex) {
-        captureException(ex);
-        console.error(ex);
-        router.replace("/redesign/user/dashboard");
-      }
-    }
-    void updateSession();
-    // Only run once, since we will be immediately redirecting no matter what.
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
-
-  return <div id="loading">{l10n.getString("subscription-check-loading")}</div>;
-}
